Impl weapon projectile intial velocity computation as an extension method to separate interface from impl
