+++ crates/thetawave_interface/Cargo.toml
@@ -17,4 +17,3 @@ leafwing-input-manager = { workspace = true }
 bevy_input = { workspace = true }
 bevy_time = { workspace = true }
 bevy_math = { workspace = true }
-rand = { workspace = true }
