+++ src/weapon/mod.rs
@@ -6,6 +6,7 @@ use bevy::{
         schedule::{common_conditions::in_state, IntoSystemConfigs},
         system::{Query, Res},
     },
+    math::Vec2,
     time::Time,
     transform::components::Transform,
 };
@@ -13,10 +14,11 @@ use bevy_rapier2d::dynamics::Velocity;
 use std::time::Duration;
 use thetawave_interface::{
     states::{AppStates, GameStates},
-    weapon::{FireMode, WeaponComponent, WeaponProjectileData},
+    weapon::{FireMode, SpreadPattern, WeaponComponent, WeaponProjectileData},
 };
 
 use crate::spawnable::{FireWeaponEvent, InitialMotion};
+use rand::{thread_rng, Rng};
 
 pub struct WeaponPlugin;
 
@@ -31,11 +33,10 @@ impl Plugin for WeaponPlugin {
     }
 }
 trait WeaponExt {
+    /// Updates the weapon's timers. Returns Some iff the weapon can be fired
     fn update(&mut self, delta_time: Duration) -> Option<WeaponProjectileData>;
 }
 impl WeaponExt for WeaponComponent {
-    /// Updates the weapon's timers
-    /// Returns true if the weapon can be fired
     fn update(&mut self, delta_time: Duration) -> Option<WeaponProjectileData> {
         if self.is_enabled {
             // tick the initial timer if there is still time remaining
@@ -79,6 +80,66 @@ pub fn update_weapon_system(
                 source_entity: entity,
                 initial_motion,
             })
+        }
+    }
+}
+
+pub(crate) trait WeaponProjectileInitialVelocitiesExt {
+    /// The initial velocities of `n` projectiles using existing/'partially evaluated' params.
+    /// Could be evenly spaced, or something else based on the struct params.
+    fn get_linvels(&self, max_projectiles: f32) -> Vec<Vec2>;
+}
+impl WeaponProjectileInitialVelocitiesExt for WeaponProjectileData {
+    fn get_linvels(&self, max_projectiles: f32) -> Vec<Vec2> {
+        match &self.spread_pattern {
+            SpreadPattern::Arc(arc_pattern) => {
+                // Get the segment of a spread angle
+                let spread_angle_segment = {
+                    // percentage of the game's maximum amount of projectiles being spawned
+                    let total_projectiles_percent =
+                        (self.count as f32 - 1.) / (max_projectiles - 1.);
+                    // indicates the angle between the first and last projectile
+                    let spread_arc = arc_pattern
+                        .max_spread
+                        .min(total_projectiles_percent * arc_pattern.projectile_gap);
+                    // indicates the angle between each projectile
+                    spread_arc / (self.count as f32 - 1.).max(1.)
+                };
+
+                let mut linvels = vec![];
+
+                for p in 0..self.count {
+                    // Calculate the angle for the current projectile.
+                    // The first projectile is spread_angle_segment/2 radians to the left of the direction,
+                    // and the last projectile is spread_angle_segment/2 radians to the right.
+                    let angle_offset =
+                        (p as f32 - (self.count as f32 - 1.) / 2.) * spread_angle_segment;
+                    let projectile_angle = self.direction + angle_offset;
+
+                    linvels.push(
+                        Vec2::from_angle(projectile_angle)
+                            * self.speed
+                            * arc_pattern.spread_weights,
+                    );
+                }
+
+                linvels
+            }
+            SpreadPattern::Random(random_pattern) => {
+                let mut linvels = vec![];
+
+                for _ in 0..self.count {
+                    linvels.push(
+                        // multiply the speed the projectile by a random angle and velocity multiplier
+                        Vec2::from_angle(
+                            thread_rng().gen_range(random_pattern.angle_range.clone()),
+                        ) * self.speed
+                            * thread_rng().gen_range(random_pattern.speed_range.clone()),
+                    );
+                }
+
+                linvels
+            }
         }
     }
 }
